Add selectEntireGrid utility function and refactor selection logic in context menu and copy keybind plugins
